Add inventory/spell tracking...
If the player has the spell "Light", he can freely toggle the torch.

If the player has a torch, and they toggle it, a torch will be used.

If the player does not know Light or have a torch, they cannot toggle.

There's a new setting to control if the GM toggling the torch uses the
actors inventory.